Refactor: split tools into chrome-tools, summon-tools, browser-tools
Split the monolithic browser-tools.ts into focused modules:

- chrome-tools.ts: Chrome-specific (switchTab, openUrl)
- summon-tools.ts: Phone/Zoom audio sync (play/pause/resume video,
  summon, dismiss, join_zoom, join_gmeet, lookup_meeting_id)
- browser-tools.ts: General computer use (scroll, describe_screen,
  capture_screen, type_text, click, scroll_and_describe, open_file,
  screen_record + all recording infrastructure)
- inline-tools.ts: Non-browser tools + registration arrays

Changes:
- scroll: now detects frontmost app — Chrome uses JS scrollBy,
  other apps use Page Down/Up keyboard events
- openVideo → openFile: generalized to open any file
- Removed: replayVideoTool (identical to playVideoTool),
  closeVideoTool (press_key cmd+q replaces it)
- All tools remain registered in inlineTools + ownerOnlyTools arrays
